LinkGenerator: Interface package rename
A mistake was made during refactoring. This fixes the mistake
