+++ link-generator/src/main/java/echoic/linkgenerator/config/LinkGeneratorConfig.java
@@ -5,13 +5,13 @@ import com.wrapper.spotify.exceptions.SpotifyWebApiException;
 import com.wrapper.spotify.model_objects.credentials.ClientCredentials;
 import com.wrapper.spotify.requests.authorization.client_credentials.ClientCredentialsRequest;
 import echoic.linkgenerator.core.SpotifySearcherConverter;
-import echoic.linkgenerator.core.unittests.AgnosticLinkGenerator;
+import echoic.linkgenerator.core.interfaces.AgnosticLinkGenerator;
 //import echoic.linkgenerator.core.AppleAgnosticLinkGenerator;
 import echoic.linkgenerator.core.SpotifyAgnosticLinkGenerator;
 import echoic.linkgenerator.core.UrlTracker;
-import echoic.linkgenerator.core.unittests.ExternalAgnosticMusicEntityGenerator;
-import echoic.linkgenerator.core.unittests.ExternalTrackedUrlGenerator;
-import echoic.linkgenerator.core.unittests.SearcherConverter;
+import echoic.linkgenerator.core.interfaces.ExternalAgnosticMusicEntityGenerator;
+import echoic.linkgenerator.core.interfaces.ExternalTrackedUrlGenerator;
+import echoic.linkgenerator.core.interfaces.SearcherConverter;
 import echoic.linkgenerator.external.linktracking.TinyCcExternalTrackedUrlGenerator;
 import echoic.linkgenerator.external.musicentity.SonglinkExtAgnosticMusicEntityGenerator;
 import lombok.extern.slf4j.Slf4j;
