+++ link-generator/src/main/java/echoic/linkgenerator/core/interfaces/AgnosticLinkGenerator.java
@@ -1,4 +1,4 @@
-package echoic.linkgenerator.core.unittests;
+package echoic.linkgenerator.core.interfaces;
 
 import echoic.linkgenerator.core.MusicEntity;
