+++ link-generator/src/main/java/echoic/linkgenerator/core/interfaces/ExternalTrackedUrlGenerator.java
@@ -1,4 +1,4 @@
-package echoic.linkgenerator.core.unittests;
+package echoic.linkgenerator.core.interfaces;
 
 public interface ExternalTrackedUrlGenerator extends TrackedUrlGenerator, ExternalGenerator
 {
