+++ link-generator/src/main/java/echoic/linkgenerator/core/interfaces/FullGenerator.java
@@ -1,4 +1,4 @@
-package echoic.linkgenerator.core.unittests;
+package echoic.linkgenerator.core.interfaces;
 
 public interface FullGenerator extends AgnosticLinkGenerator, TrackedUrlGenerator
 {
