+++ link-generator/src/main/java/echoic/linkgenerator/core/interfaces/SearcherConverter.java
@@ -1,4 +1,4 @@
-package echoic.linkgenerator.core.unittests;
+package echoic.linkgenerator.core.interfaces;
 
 public interface SearcherConverter<T> extends Searcher<T>, MusicEntityConverter<T>
 {
