+++ link-generator/src/main/java/echoic/linkgenerator/external/linktracking/TinyCcExternalTrackedUrlGenerator.java
@@ -1,6 +1,6 @@
 package echoic.linkgenerator.external.linktracking;
 
-import echoic.linkgenerator.core.unittests.ExternalTrackedUrlGenerator;
+import echoic.linkgenerator.core.interfaces.ExternalTrackedUrlGenerator;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.*;
 import org.springframework.web.client.RestTemplate;
