+++ link-generator/src/test/java/echoic/linkgenerator/core/unittests/ExternalAgnosticMusicEntityGenerator_UnitTests.java
@@ -1,5 +1,6 @@
 package echoic.linkgenerator.core.unittests;
 
+import echoic.linkgenerator.core.interfaces.ExternalAgnosticMusicEntityGenerator;
 import echoic.linkgenerator.external.ExternalAgnosticMusicEntity;
 import echoic.linkgenerator.external.Platforms;
 import org.junit.jupiter.api.BeforeAll;
