+++ link-generator/src/test/java/echoic/linkgenerator/core/unittests/SearcherConverter_UnitTests.java
@@ -6,6 +6,9 @@ import com.wrapper.spotify.model_objects.specification.Track;
 import com.wrapper.spotify.requests.authorization.client_credentials.ClientCredentialsRequest;
 import com.wrapper.spotify.requests.data.search.simplified.SearchTracksRequest;
 import echoic.linkgenerator.core.MusicEntity;
+import echoic.linkgenerator.core.interfaces.ExternalAgnosticMusicEntityGenerator;
+import echoic.linkgenerator.core.interfaces.ExternalTrackedUrlGenerator;
+import echoic.linkgenerator.core.interfaces.SearcherConverter;
 import echoic.linkgenerator.external.ExternalAgnosticMusicEntity;
 import echoic.linkgenerator.external.linktracking.TrackedUrl;
 import org.junit.jupiter.api.*;
